+++ .github/workflows/test.yml
@@ -23,4 +23,4 @@ jobs:
         run: cargo build
 
       - name: Run tests
-        run: cargo test
+        run: cargo test -- --test-threads=1
